Number theoretic problem
Number theoretic problem. Try to find the smallest integer such that when you move the first cipher of the integer and put it behind the last digit of your number the number becomes 3 times as smaller. 
NOTE: For a computer with massive computing power, it might be able to solve this, but i've gone as far as looping up to 10^9 and found no solutions. If you come up with a better way to solve this problem, please message me!